+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 ### unreleased
+- **fix:** rename Wurst.factory into Wurst.init
 
 ### [v0.4.2] - 2016-05-19
 - **add:** function to check if arguments passed to oolo are objects